Update old streams
