+++ include/exi/Stream/Stream.hpp
@@ -24,6 +24,8 @@
 #pragma once
 
 #include <core/Common/ArrayRef.hpp>
+#include <core/Common/Fundamental.hpp>
+#include <core/Common/CRTPTraits.hpp>
 #include <core/Support/ErrorHandle.hpp>
 #include <core/Support/Limits.hpp>
 #include <exi/Basic/ErrorCodes.hpp>
@@ -59,22 +61,37 @@ using ref_or_value_t = typename H::RefOrValue<T>::type;
 // BitStream*
 //======================================================================//
 
-template <class BufferT> class BitStreamCommon;
+class ReaderBase;
+class OrderedReader;
+class ChannelReader;
 
-class BitStreamReader;
-class BitStreamWriter;
+class BitReader;
+class ByteReader;
+class BlockReader;
+// TODO: DeflateReader
 
-class ByteStreamReader;
-class ByteStreamWriter;
+class WriterBase;
+
+class BitWriter;
+class ByteWriter;
+class BlockWriter;
+// TODO: DeflateWriter
 
 /// The base for BitStream types. Provides common definitions.
 struct StreamBase {
-  using size_type = u64;
-  using WordType  = u64;
+  enum StreamKind {
+    SK_Bit,     // Bit Packed
+    SK_Byte,    // Byte Packed
+    SK_Block,   // Precompression
+    SK_Deflate, // Compression
+  };
+
+  using size_type = usize;
+  using word_t    = u64;
+  using WordType  = word_t;
 
   static constexpr size_type kWordSize = sizeof(WordType);
   static constexpr size_type kBitsPerWord = bitsizeof_v<WordType>;
-  static constexpr size_type kOctetsPerWord = kBitsPerWord / 8;
 
   static constexpr size_type kMaxCapacityBytes = max_v<size_type> / kCHAR_BIT;
   static constexpr size_type kMask = (kCHAR_BIT - 1ull);
@@ -82,176 +99,13 @@ struct StreamBase {
 
 /// A proxy type for passing around consumed bits. Useful when swapping
 /// between stream types (generally between the header and body).
-template <class BufferT> class BitConsumerProxy {
-  template <typename> friend class BitStreamCommon;
+template <class BufferT> class StreamProxy {
+public:
   BufferT Bytes;
-  StreamBase::WordType Bits;
+  StreamBase::word_t NBits;
 public:
-  BitConsumerProxy(BufferT Bytes, u64 Bits) :
-   Bytes(Bytes), Bits(Bits) {
-  }
-};
-
-/// The interface for BitStream types. Provides a simple interface for reading
-/// the current position in bits and bytes, and wraps a "stream" buffer.
-template <class BufferT> class BitStreamCommon : public StreamBase {
-  static_assert(sizeof(typename BufferT::value_type) == 1,
-    "BufferT::value_type must be bytes!");
-  static_assert(std::is_pointer_v<typename BufferT::iterator>,
-    "BufferT::iterator must be a pointer!");
-public:
-  using ProxyT = BitConsumerProxy<BufferT>;
-  using value_type = typename BufferT::value_type;
-  using pointer = typename BufferT::iterator;
-  using const_pointer = typename BufferT::const_iterator;
-  using ref = ref_or_value_t<pointer>;
-  using const_ref = ref_or_value_t<const_pointer>;
-
-protected:
-  BufferT Stream;
-  // StreamBase::WordType Store = 0;
-  size_type BitCapacity = 0;
-  size_type Position = 0;
-
-public:
-  void skip(i64 Bits) {
-    exi_assert(Bits >= 0);
-    Position += Bits;
-  }
-  void skipBytes(i64 Bytes) {
-    exi_assert(Bytes >= 0);
-    Position += (Bytes * kCHAR_BIT);
-  }
-
-  /// The overall offset in bits.
-  EXI_INLINE size_type bitPos() const { return Position; }
-  /// The overall offset in bytes, clipped.
-  size_type bytePos() const { return (Position / kCHAR_BIT); }
-  /// The overall offset in bits, clipped.
-  size_type byteBitPos() const { return (Position & ~kMask); }
-  /// The offset from the start of the current byte in bits.
-  size_type bitOffset() const { return (Position & kMask); }
-  /// The offset from the next byte in bits.
-  size_type farBitOffset() const { return kCHAR_BIT - bitOffset(); }
-  /// The offset from the next unaligned byte in bits.
-  size_type farBitOffsetInclusive() const { return (farBitOffset() & kMask); }
-
-  /// The capacity in bits.
-  size_type capacity() const { return BitCapacity; }
-  size_type capacityInBytes() const { return Stream.size(); }
-
-  /// The remaining capacity in bits.
-  size_type space() const { return capacity() - bitPos(); }
-  /// The remaining capacity in bytes.
-  size_type spaceInBytes() const { return capacityInBytes() - bytePos(); }
-
-  explicit operator bool() const { return this->isFull(); }
-  /// Checks if the current position is past the capacity.
-  EXI_INLINE bool isFull() const { return bytePos() >= capacityInBytes(); }
-  /// Checks if the current position is NOT past the capacity.
-  EXI_INLINE bool notFull() const { return bytePos() < capacityInBytes(); }
-  /// Checks if the current position is byte aligned.
-  bool isByteAligned() const { return bitOffset() == 0; }
-
-  void setStream(BufferT NewStream) {
-    exi_assert(NewStream.size() <= kMaxCapacityBytes,
-      "Stream size exceeds max capacity.");
-    this->Stream = NewStream;
-    this->BitCapacity = NewStream.size() * kCHAR_BIT;
-    this->Position = 0;
-  }
-
-  ProxyT getProxy() const {
-    return ProxyT(Stream, Position);
-  }
-
-  template <typename AnyT>
-  void setProxy(BitConsumerProxy<AnyT> Other) {
-    this->setStream(Other.Bytes);
-    this->Position = Other.Bits;
-  }
-
-protected:
-  BitStreamCommon(BufferT Stream) :
-   Stream(Stream), BitCapacity(Stream.size() * kCHAR_BIT) {
-    exi_assert(Stream.size() <= kMaxCapacityBytes,
-      "Stream size exceeds max capacity.");
-  }
-
-  template <typename AnyT>
-  BitStreamCommon(BitConsumerProxy<AnyT> Other) :
-   BitStreamCommon(Other.Bytes) {
-    this->Position = Other.Bits;
-  }
-
-  ExiError ec() const noexcept {
-    if EXI_UNLIKELY(isFull())
-      return ExiError::FULL;
-    return ExiError::OK;
-  }
-
-  /// Aligns stream up to the next byte.
-  /// @returns `false` if capacity is reached, `true` otherwise.
-  template <bool CheckFull = false> bool align() {
-    if (!isByteAligned())
-      Position += (kCHAR_BIT - bitOffset());
-    if constexpr (CheckFull)
-      return EXI_LIKELY(notFull());
-    return true;
-  }
-
-  /// Aligns stream down to the current byte.
-  /// @returns `false` if capacity was reached, `true` otherwise.
-  template <bool CheckFull = false> bool alignDown() {
-    this->Position = byteBitPos();
-    if constexpr (CheckFull)
-      return EXI_LIKELY(notFull());
-    return true;
-  }
-
-  pointer getCurrentBytePtr() {
-    return &currentByteRef();
-  }
-  const_pointer getCurrentBytePtr() const {
-    return &currentByteRef();
-  }
-
-  ref getCurrentByte() {
-    return currentByteRef();
-  }
-  const_ref getCurrentByte() const {
-    return currentByteRef();
-  }
-
-  /// Check if `N` bits can be read.
-  inline bool canAccessBits(i64 N) const {
-    if EXI_UNLIKELY(N < 0)
-      return false;
-    return EXI_LIKELY(Position + N <= capacity());
-  }
-
-  /// Check if `N` bytes can be read.
-  inline bool canAccessBytes(i64 N) const {
-    return canAccessBits(N * kCHAR_BIT);
-  }
-
-  /// Check if `N` "words" can be read.
-  bool canAccessWords(i64 N = 1) const {
-    if EXI_UNLIKELY(N < 0)
-      return false;
-    const size_type Pos = (Position + 7) / kCHAR_BIT;
-    const size_type Offset = Pos + (N * kWordSize);
-    return Offset <= capacityInBytes();
-  }
-
-private:
-  ALWAYS_INLINE decltype(auto) currentByteRef() {
-    exi_invariant(!isFull());
-    return Stream[bytePos()];
-  }
-  ALWAYS_INLINE decltype(auto) currentByteRef() const {
-    exi_invariant(!isFull());
-    return Stream[bytePos()];
+  StreamProxy(BufferT Bytes, u64 Bits) :
+   Bytes(Bytes), NBits(Bits) {
   }
 };
