+++ include/exi/StreamOld/BitStreamReader.hpp
