+++ include/exi/StreamOld/BitStreamWriter.hpp
