+++ include/exi/StreamOld/StreamVariant.hpp
