+++ lib/exi/StreamOld/BitStreamCommon.hpp
